backend/requirements/requirements-py3dev.txt
Parsley==1.2
Twisted==17.9.0
acme==0.22.2
cffi==1.11.5
chardet==2.3.0
cryptography==2.1.4
idna==2.6
pyOpenSSL==17.5.0
pyasn1==0.1.9
pyasn1_modules==0.0.7
pycparser==2.14
python_debian==0.1.28
python_gnupg==0.4.2
requests==2.9.1
scrypt==0.6.1
service_identity==16.0.0
six==1.10.0
sqlalchemy==1.1.11
txtorcon==0.20.0
zope.component==4.2.2
zope.event==4.2.0
zope.interface==4.1.3
